Menu: extract the Communications tab into build_communications.cpp
Verbatim move of the LoRa menu, Phone (KDE Connect) item, notification
log and scanned-QR browser. The Phone item still lands under System >
Connectivity, handed over through ctx.phone_item; push_notif comes from
the shared context helper.
